Use componentShouldUpdate and useMemo.
In Class Component, we can use componentShouldUpdate when prevent useless update.
But in Function Component, we can't use componentShouldUpdate.
So we need to use useMemo instead of componentShouldUpdate.
It compares previous properties with new properties that are passed from Parents.
If React finds difference between them, it renders the component.