Use React-Router-Dom
Use React-Router-Dom to  make Single Page Application.
First We need to wrap the components which need to use Route,
by Router.
Second We need to Link for moving specific Url.
Third We need to use Route for rendering page which you want to see
in specific Url.

Component can recieve properties (histroy, location, match), if it is
contained in Route.
But it doesn't send to it's child. So you need to send the properties to
its children.
But if you use withRouter, high-order-component, you don't have to
send properties to its child.